+++ src/Recipies/03_TenantRouting.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ProxyKit.Examples
+namespace ProxyKit.Recipies
 {
     public class MultiTenantJwt : ExampleBase<MultiTenantJwt.Startup>
     {
