+++ src/Recipies/06_Testing.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,7 +8,7 @@ using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using ProxyKit.Testing;
 
-namespace ProxyKit.Examples
+namespace ProxyKit.Recipies
 {
     public class Testing
     {
